rebase api
