+++ app/src/main/java/com/example/lsdchat/api/ApiManager.java
@@ -1,12 +1,12 @@
 package com.example.lsdchat.api;
 
-import com.example.lsdchat.api.request.LoginRequest;
-import com.example.lsdchat.api.request.SessionRequestAuth;
-import com.example.lsdchat.api.request.SessionRequestNoAuth;
-import com.example.lsdchat.api.response.LoginResponse;
-import com.example.lsdchat.api.response.SessionResponse;
-import com.example.lsdchat.api.service.LoginService;
-import com.example.lsdchat.api.service.SessionService;
+import com.example.lsdchat.api.login.request.LoginRequest;
+import com.example.lsdchat.api.login.request.SessionRequestAuth;
+import com.example.lsdchat.api.login.request.SessionRequestNoAuth;
+import com.example.lsdchat.api.login.response.LoginResponse;
+import com.example.lsdchat.api.login.response.SessionResponse;
+import com.example.lsdchat.api.login.service.LoginService;
+import com.example.lsdchat.api.login.service.SessionService;
 import com.example.lsdchat.constant.ApiConstant;
 import com.google.gson.GsonBuilder;
