+++ app/src/main/java/com/example/lsdchat/api/login/model/LoginUser.java
@@ -1,4 +1,4 @@
-package com.example.lsdchat.api.model;
+package com.example.lsdchat.api.login.model;
 
 
 import com.google.gson.annotations.SerializedName;
