+++ app/src/main/java/com/example/lsdchat/api/login/model/SessionToken.java
@@ -1,8 +1,6 @@
-package com.example.lsdchat.api.model;
+package com.example.lsdchat.api.login.model;
 
 import com.google.gson.annotations.SerializedName;
-
-import io.realm.RealmObject;
 
 
 public class SessionToken {
