+++ app/src/main/java/com/example/lsdchat/api/login/request/SessionRequestNoAuth.java
@@ -1,9 +1,9 @@
-package com.example.lsdchat.api.request;
+package com.example.lsdchat.api.login.request;
 
 import com.google.gson.annotations.SerializedName;
 
 
-public class SessionRequestAuth {
+public class SessionRequestNoAuth {
     @SerializedName("application_id")
     private String applicationId;
     @SerializedName("auth_key")
@@ -14,20 +14,16 @@ public class SessionRequestAuth {
     private long timestamp;
     @SerializedName("signature")
     private String signature;
-    @SerializedName("user[email]")
-    private String email;
-    @SerializedName("user[password]")
-    private String password;
 
-    public SessionRequestAuth(String applicationId, String authKey, int nonce, long timestamp, String signature, String email, String password) {
+    public SessionRequestNoAuth(String applicationId, String authKey, int nonce, long timestamp, String signature) {
         this.applicationId = applicationId;
         this.authKey = authKey;
         this.nonce = nonce;
         this.timestamp = timestamp;
         this.signature = signature;
-        this.email = email;
-        this.password = password;
     }
+
+
 
 
 }
