+++ app/src/main/java/com/example/lsdchat/api/login/service/SessionService.java
@@ -1,7 +1,7 @@
-package com.example.lsdchat.api.service;
+package com.example.lsdchat.api.login.service;
 
-import com.example.lsdchat.api.request.SessionRequestNoAuth;
-import com.example.lsdchat.api.response.SessionResponse;
+import com.example.lsdchat.api.login.request.SessionRequestNoAuth;
+import com.example.lsdchat.api.login.response.SessionResponse;
 import com.example.lsdchat.constant.ApiConstant;
 
 import retrofit2.http.Body;
@@ -12,7 +12,7 @@ import rx.Observable;
 
 public interface SessionService {
 
-    @Headers("Content-Type: application/json")
+    @Headers(ApiConstant.HEADER_CONTENT_TYPE)
     @POST(ApiConstant.SESSION_REQUEST)
     Observable<SessionResponse> getSession(@Body SessionRequestNoAuth body);
