--- app/src/main/java/com/example/lsdchat/api/request/SessionRequestNoAuth.java
@@ -1,77 +1,0 @@
-package com.example.lsdchat.api.request;
-
-import com.example.lsdchat.constant.ApiConstant;
-import com.example.lsdchat.util.SignatureTest;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.Random;
-
-
-
-public class SessionRequestNoAuth {
-    @SerializedName("application_id")
-    private String applicationId;
-    @SerializedName("auth_key")
-    private String authKey;
-    @SerializedName("nonce")
-    private int nonce;
-    @SerializedName("timestamp")
-    private long timestamp;
-    @SerializedName("signature")
-    private String signature;
-
-    public SessionRequestNoAuth(String applicationId, String authKey, int nonce, long timestamp, String signature) {
-        this.applicationId = applicationId;
-        this.authKey = authKey;
-        this.nonce = nonce;
-        this.timestamp = timestamp;
-        this.signature = signature;
-    }
-
-
-    public String getApplicationId() {
-        return applicationId;
-    }
-
-    public void setApplicationId(String applicationId) {
-        this.applicationId = applicationId;
-    }
-
-    public String getAuthKey() {
-        return authKey;
-    }
-
-    public void setAuthKey(String authKey) {
-        this.authKey = authKey;
-    }
-
-    public int getNonce() {
-        return nonce;
-    }
-
-    public void setNonce(int nonce) {
-        this.nonce = nonce;
-    }
-
-    public long getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(long timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public String getSignature() {
-        return signature;
-    }
-
-    public void setSignature(String signature) {
-        this.signature = signature;
-    }
-
-
-
-
-
-
-}
